+++ Hw4/src/main/java/hw4/hw4/Service/RaceService.java
@@ -7,9 +7,14 @@ import hw4.hw4.DTO.Race.RaceDTO_PilotStatistic_CountryUSA;
 import hw4.hw4.Entity.Pilot;
 import hw4.hw4.Entity.Race;
 import hw4.hw4.Entity.RacePilot.RacesPilots;
+import hw4.hw4.Entity.User.ERole;
+import hw4.hw4.Entity.User.User;
 import hw4.hw4.Exception.RaceNotFoundException;
+import hw4.hw4.Exception.UserNotAuthorizedException;
+import hw4.hw4.Exception.UserNotFoundException;
 import hw4.hw4.Repository.RaceRepository;
 import hw4.hw4.Repository.RacesPilotsRepository;
+import hw4.hw4.Repository.UserRepository;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -26,10 +31,12 @@ public class RaceService {
     private final RaceRepository raceRepository;
     private final RacesPilotsRepository racesPilotsRepository;
 
-    public RaceService(RaceRepository raceRepository, RacesPilotsRepository racesPilotsRepository)
-    {
+    private final UserRepository userRepository;
+
+    public RaceService(RaceRepository raceRepository, RacesPilotsRepository racesPilotsRepository, UserRepository userRepository) {
         this.raceRepository = raceRepository;
         this.racesPilotsRepository = racesPilotsRepository;
+        this.userRepository = userRepository;
     }
 
     public List<RaceDTO_All> getAllRaces(Integer pageNo, Integer pageSize) {
@@ -39,6 +46,10 @@ public class RaceService {
                 race -> RaceDTO_Converters.convertToRaceDTO_All(race,
                         this.racesPilotsRepository.countByRaceId(race.getId()))
         ).collect(Collectors.toList());
+    }
+
+    public Long getRacesCount() {
+        return this.raceRepository.count();
     }
 
     public Race getOneRace(Long id) {
@@ -61,34 +72,47 @@ public class RaceService {
     }
 
     public List<RaceDTO_PilotStatistic_CountryUSA> getRacesFromUSAWithNumberOfPilotsDesc() {
+
         return this.raceRepository.getRacesFromUSAWithNumberOfPilotsDesc().stream().limit(100).collect(Collectors.toList());
+
     }
 
-    public Long getRacesCount() {
-        return this.raceRepository.count();
-    }
+    public Race addRace(Race newRace, Long userId) {
+        User user = this.userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(userId));
+        newRace.setUser(user);
 
-    public Race addRace(Race newRace) {
         return raceRepository.save(newRace);
     }
 
-    public Race updateRace(Race newRace, Long id) {
-        return raceRepository.findById(id)
-                .map(race -> {
-                    race.setName(newRace.getName());
-                    race.setDate(newRace.getDate());
-                    race.setNumberOfLaps(newRace.getNumberOfLaps());
-                    race.setCountry(newRace.getCountry());
-                    race.setLapLength(newRace.getLapLength());
-                    return raceRepository.save(race);
-                })
-                .orElseGet(() -> {
-                    newRace.setId(id);
-                    return raceRepository.save(newRace);
-                });
+    public Race updateRace(Race newRace, Long raceID, Long userID) {
+        Race race = this.raceRepository.findById(raceID).orElseThrow(() -> new RaceNotFoundException(raceID));
+        User user = this.userRepository.findById(userID).orElseThrow(() -> new UserNotFoundException(userID));
+
+        if (!Objects.equals(user.getId(), race.getUser().getId())) {
+            boolean modOrAdmin = user.getRoles().stream().anyMatch((role) ->
+                    role.getName() == ERole.ROLE_ADMIN || role.getName() == ERole.ROLE_MODERATOR
+            );
+
+            if (!modOrAdmin) {
+                throw new UserNotAuthorizedException(String.format("%s does not have permission to " +
+                        "update race %s", user.getUsername(), race.getId()));
+            }
+        }
+
+        return raceRepository.findById(raceID)
+                .map(raceUpdate -> {
+                    raceUpdate.setName(newRace.getName());
+                    raceUpdate.setDate(newRace.getDate());
+                    raceUpdate.setNumberOfLaps(newRace.getNumberOfLaps());
+                    raceUpdate.setCountry(newRace.getCountry());
+                    raceUpdate.setLapLength(newRace.getLapLength());
+                    return raceRepository.save(raceUpdate);
+                }).orElseThrow(() -> new RaceNotFoundException(raceID));
     }
 
     public void deleteRace(Long id) {
+        if(!raceRepository.existsById(id))
+            throw new RaceNotFoundException(id);
         raceRepository.deleteById(id);
     }
